// src/api/generated/model/createGrantRequestScopeType.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CreateGrantRequestScopeType = typeof CreateGrantRequestScopeType[keyof typeof CreateGrantRequestScopeType];


export const CreateGrantRequestScopeType = {
  NAMESPACE: 'NAMESPACE',
  DATABASE: 'DATABASE',
} as const;